pyproject.toml
[project]
name = "rag-question-answering"
version = "0.1.0"
description = "RAG question-answering system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.5.5",
    "gradio>=6.9.0",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "litellm>=1.82.4",
    "numpy>=2.4.3",
    "openai>=2.29.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.3.0",
    "tenacity>=9.1.4",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"